% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortSourceCodeCheck.R
\name{checkCohortSourceCodes}
\alias{checkCohortSourceCodes}
\title{Check source codes used in a cohort definition}
\usage{
checkCohortSourceCodes(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = NULL, cohortJson, cohortSql, outputFile)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{cohortJson}{A characteric string containing the JSON of a cohort definition.}

\item{cohortSql}{The OHDSI SQL representation of the same cohort definition.}

\item{outputFile}{The name of the HTML file to create.}
}
\description{
This function first extracts all concept sets used in a cohort definition. Then, for each concept set
the concept found in the CDM database the contributing source codes are identified. OVerall and per month counts
are computed and shown for all concept sets, concepts, and source codes in a HTML table.

This function requires the rmarkdown package as well as PanDoc.
}
