% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MethodEvaluation.R
\name{computeAuc}
\alias{computeAuc}
\title{Compute the area under the ROC curve}
\usage{
computeAuc(methodResults, referenceSet, confidenceIntervals = TRUE)
}
\description{
Compute the area under the ROC curve
}

