% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{computeAucs}
\alias{computeAucs}
\title{Compute the AUCs for various injected signal sizes}
\usage{
computeAucs(logRr, trueLogRr)
}
\arguments{
\item{logRr}{A vector containing the log of the relative risk as estimated by a method.}

\item{trueLogRr}{A vector containing the injected log(relative risk) for each estimate.}
}
\value{
A data frame with per injected signal size the AUC and the 95 percent confidence interval of the AUC.
}
\description{
Compute the AUCs for various injected signal sizes
}
