% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{computeMetrics}
\alias{computeMetrics}
\title{Compute the AUC, coverage, MSE, and type 1 and 2 error}
\usage{
computeMetrics(logRr, seLogRr, trueLogRr)
}
\arguments{
\item{logRr}{A numeric vector of effect estimates on the log scale}

\item{seLogRr}{The standard error of the log of the effect estimates. Hint: often the standard
error = (log(<lower bound 95 percent confidence interval>) - log(<effect
estimate>))/qnorm(0.025)}

\item{trueLogRr}{A vector of the true effect sizes}
}
\description{
Compute the AUC, coverage, MSE, and type 1 and 2 error
}
\details{
Compute the AUC, coverage, MSE, and type 1 and 2 error.
}
