% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{computeMse}
\alias{computeMse}
\title{Compute the mean squared error}
\usage{
computeMse(logRr, trueLogRr)
}
\arguments{
\item{logRr}{A numeric vector of effect estimates on the log scale.}

\item{trueLogRr}{A vector of the true effect sizes.}
}
\description{
Compute the mean squared error
}
