% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{computeType1And2Error}
\alias{computeType1And2Error}
\title{Compute type 1 and 2 error}
\usage{
computeType1And2Error(logRr, seLogRr, trueLogRr, alpha = 0.05)
}
\arguments{
\item{logRr}{A numeric vector of effect estimates on the log scale.}

\item{seLogRr}{The standard error of the log of the effect estimates. Hint: often the standard
error = (log(<lower bound 95 percent confidence interval>) - log(<effect
estimate>))/qnorm(0.025).}

\item{trueLogRr}{A vector of the true effect sizes.}

\item{alpha}{The alpha (expected type I error).}
}
\description{
Compute type 1 and 2 error
}

