% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CreateOutcomeCohorts.R
\name{createOutcomeCohorts}
\alias{createOutcomeCohorts}
\title{Create outcomes of interest}
\usage{
createOutcomeCohorts(connectionDetails, cdmDatabaseSchema,
  createNewCohortTable = FALSE, cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort", referenceSet = "omopReferenceSet")
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{A database schema containing health care data in the OMOP Commond Data Model.
Note that for SQL Server, botth the database and schema should be specified, e.g. 'cdm_schema.dbo'}

\item{createNewCohortTable}{Should a new cohort table be created, or should the outcomes be inserted in a existing table?}

\item{cohortDatabaseSchema}{The database schema where the target table is located. Note that for SQL Server, botth the
database and schema should be specified, e.g. 'cdm_schema.dbo'}

\item{cohortTable}{The name of the table where the outcomes will be stored.}
}
\description{
Create outcomes of interest
}
\details{
This function will create the outcomes of interest referenced in the various reference sets.
The outcomes of interest are derives using information like diagnoses, procedures, and drug prescriptions.
The outcomes are stored in a table on the database server.
}

