% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateOutcomeCohorts.R
\name{createReferenceSetCohorts}
\alias{createReferenceSetCohorts}
\title{Create cohorts used in a reference set.}
\usage{
createReferenceSetCohorts(connectionDetails, oracleTempSchema = NULL,
  cdmDatabaseSchema, outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "outcomes", nestingDatabaseSchema = cdmDatabaseSchema,
  nestingTable = "nesting", referenceSet = "omopReferenceSet")
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{cdmDatabaseSchema}{A database schema containing health care data in the OMOP Commond Data
Model. Note that for SQL Server, botth the database and schema should
be specified, e.g. 'cdm_schema.dbo'}

\item{outcomeDatabaseSchema}{The database schema where the target outcome table is located. Note that for
SQL Server, both the database and schema should be specified, e.g.
'cdm_schema.dbo'}

\item{outcomeTable}{The name of the table where the outcomes will be stored.}

\item{nestingDatabaseSchema}{The database schema where the nesting outcome table is located. Note that for
SQL Server, both the database and schema should be specified, e.g.
'cdm_schema.dbo'. (For the OHDSI negative controls only)}

\item{nestingTable}{The name of the table where the nesting cohorts will be stored.
(For the OHDSI negative controls only)}

\item{referenceSet}{The name of the reference set for which outcomes need to be created.
Currently supported are "omopReferenceSet", "euadrReferenceSet", and
"ohdsiNegativeControls".}
}
\description{
Create cohorts used in a reference set.
}
\details{
This function will create the outcomes of interest and nesting cohorts referenced in the 
various reference sets. The outcomes of interest are derives using information like 
diagnoses, procedures, and drug prescriptions. The outcomes are stored in a table on 
the database server.
}
