% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodEvaluation.R
\docType{data}
\name{euadrReferenceSet}
\alias{euadrReferenceSet}
\title{The EU-ADR reference set}
\format{A data frame with 399 rows and 10 variables: \describe{ \item{exposureConceptId}{Concept ID
identifying the exposure} \item{exposureConceptName}{Name of the exposure}
\item{outcomeConceptId}{Concept ID identifying the outcome} \item{outcomeConceptName}{Name of the
outcome} \item{groundTruth}{0 = negative control, 1 = positive control}
\item{indicationConceptId}{Concept Id identifying the (primary) indication of the drug. To be used
when one wants to nest the analysis within the indication} \item{indicationConceptName}{Name of the
indication} \item{comparatorDrugConceptId}{Concept ID identifying a comparator drug that can be
used as a counterfactual} \item{comparatorDrugConceptName}{Name of the comparator drug}
\item{comparatorType}{How the comparator was selected} }}
\usage{
data(euadrReferenceSet)
}
\description{
A reference set of 43 drug-outcome pairs where we believe the drug causes the outcome (
positive controls) and 50 drug-outcome pairs where we believe the drug does not cause the 
outcome (negative controls). The controls involve 10 health outcomes of interest. Note that
originally, there was an additional positive control (Nimesulide and acute liver injury), but
Nimesulide is not in RxNorm, and is not available in many countries.
}
\references{
Coloma PM, Avillach P, Salvo F, Schuemie MJ, Ferrajolo C, Pariente A, Fourrier-Reglat A, Molokhia
M, Patadia V, van der Lei J, Sturkenboom M, Trifirò G. A reference standard for evaluation of
methods for drug safety signal detection using electronic healthcare record databases. Drug Safety
36(1):13-23, 2013
}
\keyword{datasets}

