% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrphanSourceCodes.R
\name{findOrphanSourceCodes}
\alias{findOrphanSourceCodes}
\title{Find source codes that do not roll up to their ancestor}
\usage{
findOrphanSourceCodes(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = NULL, conceptName, conceptSynonyms = NULL)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{conceptName}{The exact name of the parent concept.}

\item{conceptSynonyms}{Synonyms by which the code might be described.}
}
\value{
A data frame with orhan source codes, with counts per domain how often the code was encountered
in the CDM.
}
\description{
Find source codes that do not roll up to their ancestor
}
\details{
Searches for concepts where the name contains the concept name of interest, or any of its synonyms as provided by the user. 
Then it checks whether these concepts roll up to the main concept.
}
