% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApps.R
\name{launchMethodEvaluationApp}
\alias{launchMethodEvaluationApp}
\title{Launch the Method Evaluation Shiny app}
\usage{
launchMethodEvaluationApp(exportFolder, launch.browser = TRUE)
}
\arguments{
\item{exportFolder}{A folder where the data files for the Method Evaluation app are stored. Use
the \code{\link{packageOhdsiBenchmarkResults}} function to populate this
folder.}

\item{launch.browser}{Should the app be launched in your default browser, or in a Shiny window.
Note: copying to clipboard will not work in a Shiny window.}
}
\description{
Launch the Method Evaluation Shiny app
}
\details{
Launches a Shiny app that allows the user to explore the results.
}
