% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{packageOhdsiBenchmarkResults}
\alias{packageOhdsiBenchmarkResults}
\title{Package results of a method on the OHDSI Methods Benchmark}
\usage{
packageOhdsiBenchmarkResults(
  estimates,
  controlSummary,
  analysisRef,
  databaseName,
  exportFolder,
  referenceSet = "ohdsiMethodsBenchmark"
)
}
\arguments{
\item{estimates}{A data frame containing the estimates. See details for required columns.}

\item{controlSummary}{A data frame with the summary of the controls as generated by the
\code{\link{synthesizePositiveControls}} function.}

\item{analysisRef}{A file describing the various analyses that were performed. See details for
required columns.}

\item{databaseName}{A character string to identify the database the method was executed on.}

\item{exportFolder}{The folder where the output CSV files will written.}

\item{referenceSet}{The name of the reference set for which to package the results. Currently 
supported are "ohdsiMethodsBenchmark" and "ohdsiDevelopment".}
}
\description{
Stores the results of a method on the OHDSI Methods Benchmark in a standardized format, for example
for use in the Method Evaluation Shiny app.
}
\details{
The \code{estimates} argument should have the following columns: "targetId", "outcomeId",
"analysisId", "logRr", "seLogRr", "ci95Lb", "ci95Ub".
The \code{analysisRef} argument should have the following columns: "analysisId", "method",
"comparative", "nesting", "firstExposureOnly"
The \code{targetId} and \code{outcomeId} fields identify the specific control, and should
correspond to those in the controlSummary object.
The \code{analysisId} field is an integer that identifies a specific variant of the method. For
example, if the method is 'CohortMethod', analysisId = 1 could identify a set of settings using
propensity score matching, and analysisId = 2 could identify a set of settings using
stratification.
\code{logRr}, \code{seLogRr}, \code{ci95Lb}, and \code{ci95Ub} correspond to the log of the effect
estimate (e.g. the hazard ratio), the standard error, and the upper and lower bound of the effect
size estimate, as produced by the method.
\code{method} is a character string identifyign the method (e.g. "CohortMethod").
\code{comparative} is a boolean indicating whether the analysis can also be considerd to perform
comparative effect estimation (comparing the target to the comparator).
\code{nesting} is a boolean indicating whether the analysis is nested in the nesting cohorts
identified in the gold standard.
\code{firstExposureOnly} is a boolean indicating whether only the first exposure was used in the
analysis.
}
