% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotControls}
\alias{plotControls}
\title{Plot negative and positive control estimates.}
\usage{
plotControls(
  logRr,
  seLogRr = NULL,
  ci95Lb = NULL,
  ci95Ub = NULL,
  trueLogRr,
  estimateType = "relative risk",
  fileName = NULL,
  title = NULL
)
}
\arguments{
\item{logRr}{A numeric vector of effect estimates on the log scale.}

\item{seLogRr}{The standard error of the log of the effect estimates. Hint: often the standard
error = (log(<lower bound 95 percent confidence interval>) - log(<effect
estimate>))/qnorm(0.025). If not provided the standard error will be inferred from
the 95 percent confidence interval.}

\item{ci95Lb}{The lower bound of the 95 percent confidence interval. IF not provided it will be
inferred from the standard error.}

\item{ci95Ub}{The upper bound of the 95 percent confidence interval. IF not provided it will be
inferred from the standard error.}

\item{trueLogRr}{A vector of the true effect sizes}

\item{estimateType}{A character string to denote the effect size estimate type. Used for the x-axis
and the true effect size labels.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'. See the
function \code{ggsave} in the ggplot2 package for supported file formats.}

\item{title}{An optional title to display above the plot.}
}
\value{
A Ggplot object. Use the \code{ggsave} function to save to file.
}
\description{
Plot negative and positive control estimates.
}
