% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodEvaluation.R
\name{plotRocsInjectedSignals}
\alias{plotRocsInjectedSignals}
\title{Plot the ROC curves for various injected signal sizes}
\usage{
plotRocsInjectedSignals(logRr, trueLogRr, showAucs, fileName = NULL)
}
\arguments{
\item{logRr}{A vector containing the log of the relative risk as estimated by a method.}

\item{trueLogRr}{A vector containing the injected log(relative risk) for each estimate.}

\item{showAucs}{Should the AUCs be shown in the plot?}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'. See the
function \code{ggsave} in the ggplot2 package for supported file formats.}
}
\value{
A Ggplot object. Use the \code{ggsave} function to save to file.
}
\description{
Plot the ROC curves for various injected signal sizes
}

