% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SignalInjection.R
\name{injectSignals}
\alias{injectSignals}
\title{Inject signals in database}
\usage{
injectSignals(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema,
  exposureDatabaseSchema = cdmDatabaseSchema, exposureTable = "drug_era",
  outcomeDatabaseSchema = cdmDatabaseSchema, outcomeTable = "condition_era",
  outcomeConditionTypeConceptIds = c(), createOutputTables = TRUE,
  exposureOutcomePairs, modelType = "poisson", buildOutcomeModel = TRUE,
  firstExposureOnly = FALSE, washoutWindow = 183, riskWindowStart = 0,
  riskWindowEnd = 0, addExposureDaysToEnd = TRUE,
  firstOutcomeOnly = FALSE, effectSizes = c(1, 1.25, 1.5, 2, 4, 8),
  outputDatabaseSchema = cdmDatabaseSchema,
  outputTables = c("inj_outcomes_1", "inj_outcomes_1_25", "inj_outcomes_1_5",
  "inj_outcomes_2", "inj_outcomes_4", "inj_outcomes_8"))
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using
the function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Name of database schema that contains OMOP CDM and
vocabulary.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you
want all temporary tables to be managed. Requires
create/insert permissions to this database.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where
the exposure data used to define the exposure cohorts is
available.  If exposureTable = DRUG_ERA,
exposureDatabaseSchema is not used by assumed to be
cdmSchema.  Requires read permissions to this database.}

\item{exposureTable}{The table name that contains the exposure cohorts.  If
exposureTable <> DRUG_ERA, then expectation is exposureTable
has format of COHORT table: COHORT_DEFINITION_ID,
SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where
the data used to define the outcome cohorts is available.
If exposureTable = CONDITION_ERA, exposureDatabaseSchema is
not used by assumed to be cdmSchema.  Requires read
permissions to this database.}

\item{outcomeTable}{The table name that contains the outcome cohorts.  If
outcomeTable <> CONDITION_OCCURRENCE, then expectation is
outcomeTable has format of COHORT table:
COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{outcomeConditionTypeConceptIds}{A list of TYPE_CONCEPT_ID values that will restrict
condition occurrences.  Only applicable if outcomeTable =
CONDITION_OCCURRENCE.}

\item{createOutputTables}{Should the output tablse be created prior to inserting the
outcomes? If TRUE and the tables already exists, it will
first be deleted.}

\item{exposureOutcomePairs}{A data frame with at least two columns:
                                        \itemize{
                                          \item {"exposureConceptId" containing the drug_concept_ID
                                                or cohort_definition_id of the exposure variable}
                                          \item {"outcomeConceptId" containing the
                                                condition_concept_ID or cohort_definition_id of the
                                                outcome variable}
                                        }}

\item{modelType}{Can be either "poisson" or "survival"}

\item{buildOutcomeModel}{Should an outcome model be created to predict outcomes. New
outcomes will be inserted based on the predicted
probabilities according to this model, and this will help
preserve the observed confounding when injecting signals.}

\item{firstExposureOnly}{Should signals be injected only for the first exposure? (ie.
assuming an acute effect)}

\item{washoutWindow}{Number of days at the start of observation for which no
signals will be injected, but will be used to determine
whether exposure or outcome is the first one, and for
extracting covariates to build the outcome model.}

\item{riskWindowStart}{The start of the risk window relative to the start of the
exposure (in days). When 0, risk is assumed to start on the
first day of exposure.}

\item{riskWindowEnd}{The end of the risk window relative to the start of the
exposure. Note that typically the length of exposure is
added to this number (when the \code{addExposureDaysToEnd}
parameter is set to TRUE).}

\item{addExposureDaysToEnd}{Should length of exposure be added to the risk window?}

\item{firstOutcomeOnly}{Should only the first outcome per person be considered when
modeling the outcome?}

\item{effectSizes}{A numeric vector of effect sizes that should be inserted.}

\item{outputDatabaseSchema}{The name of the database schema that is the location of the
tables containing the new outcomesRequires write permissions
to this database.}

\item{outputTables}{A vector of table names that will contain the generated outcome
                                        cohorts. There should be one table name for each requested effect size.}
}
\value{
A data.frame listing all the drug-pairs in combination with requested effect sizes and the real inserted effect size
(might be different from the requested effect size because of sampling error).
}
\description{
Inject signals in database
}
\details{
This function will insert additional outcomes for a given set of drug-outcome pairs. It is assumed
that these drug-outcome pairs represent negative controls, so the true relative risk before
inserting any outcomes should be 1. There are two models for inserting the outcomes during the
specified risk window of the drug: a Poisson model assuming multiple outcomes could occurr during a
single exposure, and a survival model considering only one outcome per exposure.
For each
}

