% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalInjection.R
\name{injectSignals}
\alias{injectSignals}
\title{Inject signals in database}
\usage{
injectSignals(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema,
  exposureDatabaseSchema = cdmDatabaseSchema, exposureTable = "drug_era",
  outcomeDatabaseSchema = cdmDatabaseSchema, outcomeTable = "cohort",
  outputDatabaseSchema = outcomeDatabaseSchema, outputTable = outcomeTable,
  createOutputTable = FALSE, exposureOutcomePairs, modelType = "poisson",
  buildOutcomeModel = TRUE, buildModelPerExposure = FALSE,
  minOutcomeCountForModel = 100, minOutcomeCountForInjection = 25,
  covariateSettings = FeatureExtraction::createCovariateSettings(useDemographicsGender
  = TRUE, useDemographicsAge = TRUE, useDemographicsIndexYear = TRUE,
  useDemographicsIndexMonth = TRUE, useConditionOccurrenceLongTerm = TRUE,
  useConditionOccurrenceShortTerm = TRUE, useConditionEraLongTerm = TRUE,
  useConditionEraShortTerm = TRUE, useConditionGroupEraLongTerm = TRUE,
  useConditionGroupEraShortTerm = TRUE, useDrugExposureLongTerm = TRUE,
  useDrugExposureShortTerm = TRUE, useDrugEraLongTerm = TRUE,
  useDrugEraShortTerm = TRUE,      useDrugGroupEraLongTerm = TRUE,
  useDrugGroupEraShortTerm = TRUE, useProcedureOccurrenceLongTerm = TRUE,
  useProcedureOccurrenceShortTerm = TRUE, useDeviceExposureLongTerm = TRUE,
  useDeviceExposureShortTerm = TRUE, useMeasurementLongTerm = TRUE,
  useMeasurementShortTerm = TRUE, useObservationLongTerm = TRUE,
  useObservationShortTerm = TRUE, useCharlsonIndex = TRUE, longTermDays = 365,
  shortTermDays = 30, windowEndDays = 0, excludedCovariateConceptIds = c(),
  addDescendantsToExclude = FALSE, includedCovariateConceptIds = c(),     
  addDescendantsToInclude = FALSE, includedCovariateIds = c(),
  deleteCovariatesSmallCount = 100), prior = createPrior("laplace", exclude =
  0, useCrossValidation = TRUE), control = createControl(cvType = "auto",
  startingVariance = 0.1, noiseLevel = "quiet", threads = 10),
  firstExposureOnly = FALSE, washoutPeriod = 183, riskWindowStart = 0,
  riskWindowEnd = 0, addExposureDaysToEnd = TRUE,
  firstOutcomeOnly = FALSE, removePeopleWithPriorOutcomes = FALSE,
  maxSubjectsForModel = 1e+05, effectSizes = c(1, 1.25, 1.5, 2, 4),
  precision = 0.01, outputIdOffset = 1000,
  workFolder = "./SignalInjectionTemp", cdmVersion = "5",
  modelThreads = 1, generationThreads = 1)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using
the function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Name of database schema that contains OMOP CDM and
vocabulary.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you
want all temporary tables to be managed. Requires
create/insert permissions to this database.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where
the exposure data used to define the exposure cohorts is
available.  If exposureTable = DRUG_ERA,
exposureDatabaseSchema is not used by assumed to be
cdmSchema.  Requires read permissions to this database.}

\item{exposureTable}{The table name that contains the exposure cohorts.  If
exposureTable <> DRUG_ERA, then expectation is exposureTable
has format of COHORT table: cohort_concept_id,
SUBJECT_ID, COHORT_START_DATE, COHORT_END_DATE.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where
the data used to define the outcome cohorts is available.
If exposureTable = CONDITION_ERA, exposureDatabaseSchema is
not used by assumed to be cdmSchema.  Requires read
permissions to this database.}

\item{outcomeTable}{The table name that contains the outcome cohorts. When the table name is
not CONDITION_ERA This table is expected
to have the same format as the COHORT table:
SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE, COHORT_CONCEPT_ID (CDM v4) or COHORT_DEFINITION_ID (CDM v5 and higher).}

\item{outputDatabaseSchema}{The name of the database schema that is the location of the
tables containing the new outcomesRequires write permissions
to this database.}

\item{outputTable}{The name of the table names that will contain the generated outcome
cohorts.}

\item{createOutputTable}{Should the output table be created prior to inserting the
outcomes? If TRUE and the tables already exists, it will
first be deleted. If FALSE, the table is assumed to exist and the
outcomes will be inserted. Any existing outcomes with the same IDs
will first be deleted.}

\item{exposureOutcomePairs}{A data frame with at least two columns:
\itemize{
  \item {"exposureId" containing the drug_concept_ID
        or cohort_concept_id of the exposure variable}
  \item {"outcomeId" containing the
        condition_concept_ID or cohort_concept_id of the
        outcome variable}
}}

\item{modelType}{Can be either "poisson" or "survival"}

\item{buildOutcomeModel}{Should an outcome model be created to predict outcomes. New
outcomes will be inserted based on the predicted
probabilities according to this model, and this will help
preserve the observed confounding when injecting signals.}

\item{buildModelPerExposure}{If TRUE, an outcome model will be created for each exposure ID. IF false,
outcome models will be created across all exposures.}

\item{minOutcomeCountForModel}{Minimum number of outcome events required to build a model.}

\item{minOutcomeCountForInjection}{Minimum number of outcome events required to inject a signal.}

\item{covariateSettings}{An object of type \code{covariateSettings} as created using the
\code{createCovariateSettings} function in the
\code{FeatureExtraction} package.}

\item{prior}{The prior used to fit the outcome model. See \code{\link[Cyclops]{createPrior}}
for details.}

\item{control}{The control object used to control the cross-validation used to
determine the hyperparameters of the prior (if applicable). See
\code{\link[Cyclops]{createControl}} for details.}

\item{firstExposureOnly}{Should signals be injected only for the first exposure? (ie.
assuming an acute effect)}

\item{washoutPeriod}{Number of days at the start of observation for which no
signals will be injected, but will be used to determine
whether exposure or outcome is the first one, and for
extracting covariates to build the outcome model.}

\item{riskWindowStart}{The start of the risk window relative to the start of the
exposure (in days). When 0, risk is assumed to start on the
first day of exposure.}

\item{riskWindowEnd}{The end of the risk window relative to the start of the
exposure. Note that typically the length of exposure is
added to this number (when the \code{addExposureDaysToEnd}
parameter is set to TRUE).}

\item{addExposureDaysToEnd}{Should length of exposure be added to the risk window?}

\item{firstOutcomeOnly}{Should only the first outcome per person be considered when
modeling the outcome?}

\item{removePeopleWithPriorOutcomes}{Remove people with prior outcomes?}

\item{maxSubjectsForModel}{Maximum number of people used to fit an outcome model.}

\item{effectSizes}{A numeric vector of effect sizes that should be inserted.}

\item{precision}{The allowed ratio between target and injected signal size.}

\item{outputIdOffset}{What should be the first new outcome ID that is to be created?}

\item{workFolder}{Path to a folder where intermediate data will be stored.}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and "5".}

\item{modelThreads}{Number of parallel threads to use when fitting outcome models.}

\item{generationThreads}{Number of parallel threads to use when generating outcomes.}
}
\value{
A data.frame listing all the drug-pairs in combination with requested effect sizes and the real inserted effect size
(might be different from the requested effect size because of sampling error).
}
\description{
Inject signals in database
}
\details{
This function will insert additional outcomes for a given set of drug-outcome pairs. It is assumed
that these drug-outcome pairs represent negative controls, so the true relative risk before
inserting any outcomes should be 1. There are two models for inserting the outcomes during the
specified risk window of the drug: a Poisson model assuming multiple outcomes could occurr during a
single exposure, and a survival model considering only one outcome per exposure.
For each
}
