% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodEvaluation.R
\docType{package}
\name{MethodEvaluation-package}
\alias{MethodEvaluation}
\alias{MethodEvaluation-package}
\title{MethodEvaluation: Package for Evaluation of Estimation Methods}
\description{
Resources for the evaluation of the performance of methods that aim to estimate the magnitude (relative risk) of the effect of a drug on an outcome. These resources include reference sets for evaluating methods on real data, as well as functions for inserting simulated effects in real data based on negative control drug-outcome pairs. Further included are functions for the computation of the minimum detectable relative risks and functions for computing performance statistics such as predictive accuracy, error and bias.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/MethodEvaluation}
  \item Report bugs at \url{https://github.com/OHDSI/MethodEvaluation/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
