% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodEvaluation.R
\docType{data}
\name{omopReferenceSet}
\alias{omopReferenceSet}
\title{The OMOP reference set A reference set of 165 drug-outcome pairs where we believe the drug causes
the outcome ( positive controls) and 234 drug-outcome pairs where we believe the drug does not
cause the outcome (negative controls). The controls involve 4 health outcomes of interest: acute
liver injury, acute kidney injury, acute myocardial infarction, and GI bleeding.}
\format{A data frame with 399 rows and 10 variables: \describe{ \item{exposureId}{Concept ID identifying
the exposure} \item{exposureName}{Name of the exposure} \item{outcomeId}{Concept ID identifying the
outcome} \item{outcomeName}{Name of the outcome} \item{groundTruth}{0 = negative control, 1 =
positive control} \item{indicationId}{Concept Id identifying the (primary) indication of the drug.
To be used when one wants to nest the analysis within the indication} \item{indicationName}{Name of
the indication} \item{comparatorId}{Concept ID identifying a comparator drug that can be used as a
counterfactual} \item{comparatorName}{Name of the comparator drug} \item{comparatorType}{How the
comparator was selected} }}
\usage{
data(omopReferenceSet)
}
\description{
The OMOP reference set A reference set of 165 drug-outcome pairs where we believe the drug causes
the outcome ( positive controls) and 234 drug-outcome pairs where we believe the drug does not
cause the outcome (negative controls). The controls involve 4 health outcomes of interest: acute
liver injury, acute kidney injury, acute myocardial infarction, and GI bleeding.
}
\references{
Ryan PB, Schuemie MJ, Welebob E, Duke J, Valentine S, Hartzema AG. Defining a reference set to
support methodological research in drug safety. Drug Safety 36 Suppl 1:S33-47, 2013
}
\keyword{datasets}
